feat: implement BotAnimation component and integrate into main and bot tool pages with localization updates
